+++ Cargo.toml
@@ -10,5 +10,6 @@ crate-type = ["cdylib"]
 
 [dependencies]
 lazy_static = "1.5.0"
+ndarray = "0.16.1"
 numpy = "0.23.0"
 pyo3 = "0.23.3"
