+++ src/lib.rs
@@ -1,5 +1,8 @@
+mod cmaps;
 mod colorize;
-use colorize::{apply_color_map_py, apply_colors_and_merge_nc_py};
+mod interface;
+
+use interface::{apply_color_map_py, apply_colors_and_merge_nc_py};
 use pyo3::prelude::*;
 
 /// A Python module implemented in Rust.
